📝 chore(.DS_Store): add .DS_Store to .gitignore 📝 chore(README.md): update optional instructions and fix typo 📝 chore(images): add images for Code Engine job creation 📝 chore(rollingIaaS.py): update LogDNA app name
The .DS_Store file is added to the .gitignore file to prevent it from being tracked by Git.

In the README.md file, the optional instructions for deploying resources via the CLI are updated. Additionally, a typo in the file is fixed.

Images for the Code Engine job creation process are added to the images directory.

In the rollingIaaS.py file, the LogDNA app name is updated from 'refresh-job' to 'schematics-refresh'.